Add new cycleway tags
"shared" is the same as "shared_lane". There are 12,000 objects with "shared" and 7,000 objects with "shared_lane". 

"buffered_lane" is a tag that was introduced in 2013. A buffered bike lane is mostly only used in the United States. It exists where a conventional bike lane has additional width on one or both sides wherein dashed lines are added to indicate (1) that the bicyclist should ride away from them and away from parked cars' doors, and (2) to motorists more visibly that there is a bike lane that they should not cross.